+++ packages/sui-bundler/webpack.config.lib.js
@@ -1,17 +1,29 @@
 const webpack = require('webpack')
-const {cleanList, envVars, MAIN_ENTRY_POINT, config} = require('./shared/index')
+const {
+  cleanList,
+  envVars,
+  MAIN_ENTRY_POINT,
+  config
+} = require('./shared/index.js')
 const path = require('path')
-const minifyJs = require('./shared/minify-js')
-const definePlugin = require('./shared/define')
-const babelRules = require('./shared/module-rules-babel')
-const {extractComments, sourceMap} = require('./shared/config')
-const {aliasFromConfig} = require('./shared/resolve-alias')
+const minifyJs = require('./shared/minify-js.js')
+const definePlugin = require('./shared/define.js')
+const babelRules = require('./shared/module-rules-babel.js')
+const {extractComments, sourceMap} = require('./shared/config.js')
+const {aliasFromConfig} = require('./shared/resolve-alias.js')
 
 module.exports = {
   mode: 'production',
   resolve: {
     alias: {
       ...aliasFromConfig
+    },
+    fallback: {
+      assert: false,
+      fs: false,
+      http: require.resolve('stream-http'),
+      https: require.resolve('https-browserify'),
+      path: false
     },
     extensions: ['.js', '.json'],
     modules: ['node_modules', path.resolve(process.cwd())]
@@ -24,12 +36,12 @@ module.exports = {
     : MAIN_ENTRY_POINT,
   target: 'web',
   output: {
-    jsonpFunction: 'suiWebpackJsonp',
     filename: 'index.js'
   },
   optimization: {
     // avoid looping over all the modules after the compilation
     checkWasmTypes: false,
+    minimize: true,
     minimizer: [minifyJs({extractComments, sourceMap})]
   },
   plugins: cleanList([
@@ -41,10 +53,5 @@ module.exports = {
   ]),
   module: {
     rules: [babelRules]
-  },
-  node: {
-    fs: 'empty',
-    net: 'empty',
-    tls: 'empty'
   }
 }