+++ packages/sui-sass-loader/test/server/fixtures/async/index.js
@@ -1,0 +1,1 @@
+import(/* webpackChunkName: "async" */ './async.scss')